Enhancements to Usermod PIR_Sensor_Switch
I have added functionality whilst, so far as I can tell, preserving the original logic of the usermod.

I have added a Usermod toggle to turn on HomeAssistant Discovery messages for the PIR motion state and created a separate 'Occupancy' state, which runs in parallel to the motion state and is configurable in the Usermod settings.

Occupancy does nothing on the device, it is solely to be broadcast to MQTT, I did this inspired by dedicated motion sensors that handle 'occupancy' on the device, rather than requiring 3rd party to track occupancy logic.

I modified the MQTT procedure to allow passing of sensor type (motion/occupancy) and added a routine to send out discovery topics if enabled.

I have been running this in test on an ESP32 on a previous release and have updated it to reflect the changes to the upstream version in the current release.

Possible further enhancement would be to show the occupancy countdown in the info screen.